implemented matchpageB with dropdown menu
